% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pufvis.R
\name{check_crps}
\alias{check_crps}
\title{Check if CRPs are valid}
\usage{
check_crps(crps)
}
\arguments{
\item{crps}{Data frame or array containing the CRPs}
}
\value{
Nothing if the CRPs are valid. If the CRPs are not valid, an error message containing the missing columns.
}
\description{
CRPs can be provided in a data frame or in an array.
If a data frame is provided, the following columns are expected:
\itemize{
\item device, challenge, response, sample
}
}
\details{
If an array is provided, the array needs to have 3 dimensions.
}
