% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pufvis.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Calculate all metrics of the given CRP array}
\usage{
compute_metrics(crp_table, ref_sample = 1)
}
\arguments{
\item{crp_table}{A 3D array containing the PUF CRPs}

\item{ref_sample}{Numeric or character index representing the reference sample}
}
\value{
A list containing all PUF quality metrics
}
\description{
Calculate all metrics of the given CRP array
}
