% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pufvis.R
\name{df_to_crps}
\alias{df_to_crps}
\title{Convert a CRP data frame into a CRP array}
\usage{
df_to_crps(crps)
}
\arguments{
\item{crps}{Data frame containing the responses.}
}
\value{
A 3D array with the PUF CRPs
}
\description{
Convert a CRP data frame into a CRP array
}
