% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pufvis.R
\name{metrics_to_excel}
\alias{metrics_to_excel}
\title{Export PUF metrics into an XLSX file}
\usage{
metrics_to_excel(file, metrics, ...)
}
\arguments{
\item{file}{Path to the xlsx file}

\item{metrics}{The PUF quality metrics}

\item{...}{Rest of the parameters to pass to the algorithm}
}
\value{
The XLSX workbook
}
\description{
Export PUF metrics into an XLSX file
}
