% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pufvis.R
\name{pufvis}
\alias{pufvis}
\title{Open the shiny application}
\usage{
pufvis(crps = NULL, algorithms = algorithms, maxRequestSize = NULL, ...)
}
\arguments{
\item{crps}{Array or dataframe containing the crps}

\item{algorithms}{List of algorithms used to create the CPRs from the raw responses}

\item{maxRequestSize}{Maximum size in bytes of files to upload to the server}
}
\description{
Open the shiny application
}
