% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pufvis.R
\name{responses_to_crps}
\alias{responses_to_crps}
\title{Convert raw response data into CRPs}
\usage{
responses_to_crps(responses, algo, ...)
}
\arguments{
\item{responses}{Data frame or array with the raw responses}

\item{algo}{Algorithm used to generate the CRPs from the responses.}

\item{...}{Rest of the parameters to pass to the algorithm}
}
\value{
A 3D array with the PUF CRPs
}
\description{
Convert raw response data into CRPs
}
\seealso{
\link{algorithms}
}
