% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pufvis.R
\name{pufvis}
\alias{pufvis}
\title{Open the shiny application}
\usage{
pufvis(crps = NULL, algos = algorithms, ...)
}
\arguments{
\item{crps}{Array or dataframe containing the crps}

\item{...}{Rest of the arguments passed to \code{shiny::runApp}}

\item{algorithms}{List of algorithms used to create the CPRs from the raw responses}
}
\description{
Open the shiny application
}
